backend/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==1.11.15
django-js-asset==1.1.0
django-mptt==0.9.1
django-widget-tweaks==1.4.3
idna==2.6
keyring==10.6.0
Pillow==5.3.0
psycopg2==2.7.5
pycparser==2.19
pygobject==3.26.1
pytz==2018.3
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.2.4
